==> scripts/01-import-data.qmd <==
---
title: "Importar Datos"
output: html_document
---

# Setup

```{r setup , results = "hide"}
require("knitr")
opts_chunk$set(echo = TRUE)
opts_knit$set(root.dir = "..")
```

```{r, results = "hide"}
library(haven)
library(survey)
library(data.table)
library(sf)
```

Antes de trabajar con cualquier módulo de ENAHO, se recomienda leer con calma su ficha técnica y el diccionario de variables (el cual es descargado junto con los datos en el enlace provisto en la sección principal).

Luego de haber hecho ese trabajo, importaremos solo las columnas que nos interesa para responder nuestra pregunta clave: ¿a dónde va un peruano cuando se siente mal?

# Carga de los datos

```{r}
#|label : declarando_design

set.seed(123)

enaho400 <- read_dta("data/enaho01a-2023-400.dta")

modulo400 <- svydesign(id = ~conglome+vivienda,
                      strata = ~estrato,
                      weights = ~factor07,
                      data = enaho400)
```

```{r}
#| label : import_data

columnas <- c(7, 48, 50:62)                                                  # <1>
datos <- as.data.table(modulo400$variables)[, ..columnas]                    # <2>
setnames(datos, tolower(names(datos)))                                       # <3>
saveRDS(datos, file.path("outputs","datos.rds"))                             # <4>
```

1.  Selección de columnas de interés
2.  Convertir el dataset a un formato de data.table
3.  Convertir los nombres de las columnas a minúsculas
4.  Guardar la tabla en formato .rds

Veamos los datos.

```{r}
head(datos)
```

Con miras de realizar un mapa para graficar mejor la idea que queremos realizar importaremos también un un shapefile de Limites departamentales obtenido de la página de [datos abiertos del Perú](https://www.datosabiertos.gob.pe/dataset/limites-departamentales).

```{r, results = "hide"}
#| label: import shape file peru
peru_shapefile <- st_read("data/DEPARTAMENTOS.shp")                   #<1>
saveRDS(peru_shapefile, file.path("outputs", "departamentos.rds"))    #<2>
head(peru_shapefile)                                                  #<3>
```

1.  Cargamos el shapefile donde esta contenida la informacion geoespacial que necesitamos del Perú.
2.  Guardamos el archivo para su posterior uso.
3.  Visualizamos los datos.
